requirements.txt
langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.3.26
langchain-pinecone==0.2.8
pinecone-client==3.2.2
groq==0.37.1
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
pydantic==2.7.4
